requirements.txt
Django==2.2.7
djangorestframework==3.11.0
pkg-resources==0.0.0
sqlparse==0.3.0
